requirements.txt
Django
django-environ
djangorestframework
djangorestframework-simplejwt
drf-spectacular
psycopg2-binary